+++ src/main/java/jcrapi2/api/intern/locations/seasons/global/rankings/pathoflegend/TopPlayerPathOfLegendSeasonRankingsResponse.java
@@ -16,7 +16,7 @@
  */
 package jcrapi2.api.intern.locations.seasons.global.rankings.pathoflegend;
 
-import jcrapi2.common.PaginationResponse;
+import supercell.api.wrapper.essentials.common.PaginationResponse;
 
 public class TopPlayerPathOfLegendSeasonRankingsResponse
         extends PaginationResponse<TopPlayerPathOfLegendSeasonRanking> {}