+++ src/main/java/jcrapi2/api/intern/players/PlayerApi.java
@@ -16,13 +16,14 @@
  */
 package jcrapi2.api.intern.players;
 
-import jcrapi2.api.Api;
 import jcrapi2.api.intern.players.battlelog.BattleLogRequest;
 import jcrapi2.api.intern.players.battlelog.BattleLogResponse;
 import jcrapi2.api.intern.players.info.PlayerRequest;
 import jcrapi2.api.intern.players.info.PlayerResponse;
 import jcrapi2.api.intern.players.upcomingchests.UpcomingChestsRequest;
 import jcrapi2.api.intern.players.upcomingchests.UpcomingChestsResponse;
+
+import supercell.api.wrapper.essentials.api.Api;
 
 import java.util.concurrent.Future;