+++ src/main/java/jcrapi2/api/intern/players/PlayerApiImpl.java
@@ -16,14 +16,15 @@
  */
 package jcrapi2.api.intern.players;
 
-import jcrapi2.api.ApiContext;
-import jcrapi2.api.BaseApi;
 import jcrapi2.api.intern.players.battlelog.BattleLogRequest;
 import jcrapi2.api.intern.players.battlelog.BattleLogResponse;
 import jcrapi2.api.intern.players.info.PlayerRequest;
 import jcrapi2.api.intern.players.info.PlayerResponse;
 import jcrapi2.api.intern.players.upcomingchests.UpcomingChestsRequest;
 import jcrapi2.api.intern.players.upcomingchests.UpcomingChestsResponse;
+
+import supercell.api.wrapper.essentials.api.ApiContext;
+import supercell.api.wrapper.essentials.api.BaseApi;
 
 import java.util.concurrent.Future;