+++ src/test/java/jcrapi2/EndToEnd.java
@@ -28,7 +28,6 @@ import static java.util.Collections.emptyList;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 
-import jcrapi2.api.connector.FilesystemCachedConnector;
 import jcrapi2.api.intern.cards.CardApi;
 import jcrapi2.api.intern.cards.CardsRequest;
 import jcrapi2.api.intern.cards.CardsResponse;
@@ -90,6 +89,8 @@ import jcrapi2.api.intern.tournaments.info.TournamentResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
+import supercell.api.wrapper.essentials.connector.FilesystemCachedConnector;
+
 import java.io.StringReader;
 
 import javax.json.Json;
@@ -130,7 +131,7 @@ class EndToEnd {
                 new JCrApi(
                         "https://proxy.royaleapi.dev/v1",
                         getProperty("apiKey", getenv("API_KEY")),
-                        new FilesystemCachedConnector());
+                        new FilesystemCachedConnector("jcrapi2"));
         clanApi = jCrApi.getApi(ClanApi.class);
         playerApi = jCrApi.getApi(PlayerApi.class);
         cardApi = jCrApi.getApi(CardApi.class);